Huzzah, SPI works as a master!

+++ crunky/examples/bcm/bcm2835.h
@@ -686,10 +686,11 @@
 
 #include <stdlib.h>
 
+
 /*! Physical address and size of the peripherals block
   May be overridden on RPi2
 */
-extern off_t bcm2835_peripherals_base;
+extern uint32_t bcm2835_peripherals_base;
 /*! Size of the peripherals block to be mapped */
 extern size_t bcm2835_peripherals_size;
